De-spaghettified imports
