+++ libs/desktop/toasts/utils/README.md
@@ -1,0 +1,9 @@
+# desktop-toasts-utils
+
+This library was generated with [Nx](https://nx.dev).
+
+
+## Running unit tests
+
+Run `nx test desktop-toasts-utils` to execute the unit tests.
+
