+++ libs/desktop/toasts/utils/project.json
@@ -1,0 +1,38 @@
+{
+  "name": "desktop-toasts-utils",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/desktop/toasts/utils/src",
+  "prefix": "towech-finance",
+  "tags": [],
+  "projectType": "library",
+  "targets": {
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": [
+        "{workspaceRoot}/coverage/{projectRoot}"
+      ],
+      "options": {
+        "jestConfig": "libs/desktop/toasts/utils/jest.config.ts",
+        "passWithNoTests": true
+      },
+      "configurations": {
+        "ci": {
+          "ci": true,
+          "codeCoverage": true
+        }
+      }
+    },
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": [
+        "{options.outputFile}"
+      ],
+      "options": {
+        "lintFilePatterns": [
+          "libs/desktop/toasts/utils/**/*.ts",
+          "libs/desktop/toasts/utils/**/*.html"
+        ]
+      }
+    }
+  }
+}
