+++ libs/desktop/toasts/utils/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/toast-types.model';
+export * from './lib/toast.model';
