+++ libs/desktop/toasts/utils/src/lib/toast-types.model.ts
@@ -1,0 +1,6 @@
+export enum ToastTypes {
+  ACCENT = 'accent',
+  ERROR = 'error',
+  SUCCESS = 'success',
+  WARNING = 'warning',
+}
