--- libs/desktop_old/shell/ui/layout/README.md
@@ -1,7 +1,0 @@
-# desktop-shell-ui-layout
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test desktop-shell-ui-layout` to execute the unit tests.
