--- libs/desktop_old/shell/ui/layout/project.json
@@ -1,34 +1,0 @@
-{
-  "name": "desktop-shell-ui-layout",
-  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/desktop/shell/ui/layout/src",
-  "prefix": "towech-finance",
-  "tags": [],
-  "projectType": "library",
-  "targets": {
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/desktop/shell/ui/layout/jest.config.ts",
-        "passWithNoTests": true
-      },
-      "configurations": {
-        "ci": {
-          "ci": true,
-          "codeCoverage": true
-        }
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "libs/desktop/shell/ui/layout/**/*.ts",
-          "libs/desktop/shell/ui/layout/**/*.html"
-        ]
-      }
-    }
-  }
-}
