--- libs/desktop_old/shell/utils/guards/README.md
@@ -1,7 +1,0 @@
-# desktop-shell-utils-guards
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test desktop-shell-utils-guards` to execute the unit tests.
