--- libs/desktop_old/shell/utils/guards/project.json
@@ -1,34 +1,0 @@
-{
-  "name": "desktop-shell-utils-guards",
-  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/desktop/shell/utils/guards/src",
-  "prefix": "towech-finance",
-  "tags": [],
-  "projectType": "library",
-  "targets": {
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/desktop/shell/utils/guards/jest.config.ts",
-        "passWithNoTests": true
-      },
-      "configurations": {
-        "ci": {
-          "ci": true,
-          "codeCoverage": true
-        }
-      }
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "libs/desktop/shell/utils/guards/**/*.ts",
-          "libs/desktop/shell/utils/guards/**/*.html"
-        ]
-      }
-    }
-  }
-}
