--- libs/desktop_old/shell/utils/guards/src/index.ts
@@ -1,3 +1,0 @@
-export * from './lib/desktop-shell-utils-guards.module';
-export * from './lib/auth.guard';
-export * from './lib/no-auth.guard';
