--- libs/desktop_old/shell/utils/guards/src/lib/desktop-shell-utils-guards.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-import { AuthGuard } from './auth.guard';
-import { NoAuthGuard } from './no-auth.guard';
-
-@NgModule({
-  providers: [AuthGuard, NoAuthGuard],
-})
-export class DesktopShellUtilsGuardsModule {}
