--- libs/desktop_old/shell/utils/guards/src/test-setup.ts
@@ -1,1 +1,0 @@
-import 'jest-preset-angular/setup-jest';
